+++ .routify/urlIndex.json
@@ -20,7 +20,8 @@
   "/html/wknd11/practice_AppTodo",
   "/html/wknd11/practice_FruitsShop",
   "/html/wknd11/wknd11_FruitShop",
-  "/html/wknd12/wknd12_app",
+  "/html/wknd12/Circle",
+  "/html/wknd12/wknd12_main",
   "/index",
   "/test/array"
 ]
