+++ src/pages/html/wknd12/types.ts
@@ -1,0 +1,2 @@
+export type HostList = [name: string, id: string];
+export type OsInfo = [name: string, cpu: number, memory: number];
